+++ src/components/Inputs/HTTPMethodSelect/index.ts
@@ -1,0 +1,1 @@
+export { HTTPMethodSelect } from './HTTPMethodSelect'
