+++ src/components/Inputs/MillisecondInput/index.ts
@@ -1,0 +1,1 @@
+export { MillisecondInput } from './MillisecondInput'
