+++ src/components/Inputs/MillisecondInput/internal/Millisecond.typing.ts
@@ -1,0 +1,1 @@
+export type Millisecond = `${number}ms`
