+++ src/components/Inputs/NumberInput/index.ts
@@ -1,0 +1,1 @@
+export { NumberInput } from './NumberInput'
