+++ src/components/Inputs/NumberInput/internal/NumberInput.helper.ts
@@ -1,0 +1,23 @@
+import { ChangeEventHandler } from 'react'
+
+import { parseInteger } from '@/tools'
+
+export const handleChangeAsNumber = (
+  onChange: (value: number) => void
+): ChangeEventHandler<HTMLInputElement> => {
+  return (e) => onChange(getNumberValue(e.target.value))
+}
+
+export const getNumberValue = (value: string): number => {
+  if (value === '') {
+    return 0
+  }
+  return parseInteger(value)
+}
+
+export const getStringValue = (value: number | undefined): string => {
+  if (value === undefined) {
+    return ''
+  }
+  return `${value}`
+}
