+++ src/pages/Run/components/RunConfigurationPanel/components/index.ts
@@ -1,3 +1,2 @@
-export { MethodSelect } from './MethodSelect'
 export { HeadersForm } from './HeadersForm'
 export { TestsForm } from './TestsForm'
